requirements.txt
aiohttp==3.7.4.post0
aiohttp-socks==0.6.0
altair==4.1.0
appnope==0.1.2
APScheduler==3.6.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
astor==0.8.1
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.9.3
bleach==4.1.0
blinker==1.4
boto3==1.20.6
botocore==1.23.6
cachetools==4.2.2
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
crispy-bootstrap5==0.4
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
Django==3.2.6
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.4
django-generic-json-views==0.8
django-storages==1.12.3
entrypoints==0.3
fastapi==0.68.0
feedparser==6.0.8
gitdb==4.0.9
GitPython==3.1.24
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
idna==3.2
install==1.3.4
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
jmespath==0.10.0
jsonschema==4.3.1
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
mailgun2==2.0.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==5.1.0
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.6
numpy==1.21.1
packaging==21.3
pandas==1.3.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
playwright==1.20.1
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.1
psycopg2-binary==2.9.2
ptyprocess==0.7.0
pyaes==1.6.1
pyarrow==6.0.1
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
pydeck==0.7.1
pyee==8.1.0
Pygments==2.10.0
Pympler==1.0
pyparsing==3.0.6
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-socks==1.2.4
python-telegram-bot==13.9
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.3.0
requests==2.26.0
rsa==4.7.2
s3transfer==0.5.0
selenium==3.141.0
Send2Trash==1.8.0
sgmllib3k==1.0.0
six==1.10.0
smmap==5.0.0
soupsieve==2.2.1
sqlparse==0.4.1
starlette==0.14.2
streamlit==1.3.0
telebot==0.0.4
telegram==0.0.1
Telethon==1.23.0
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
toolz==0.11.2
tornado==6.1
traitlets==5.1.1
typing-extensions==3.10.0.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.6
uvicorn==0.14.0
validators==0.18.2
wcwidth==0.2.5
webdriver-manager==3.4.2
webencodings==0.5.1
websockets==10.1
whitenoise==5.3.0
widgetsnbextension==3.5.2
yarl==1.6.3
